API-34441 introduce POA request models
