+++ modules/claims_api/app/controllers/claims_api/v2/power_of_attorney_requests/base_controller.rb
@@ -2,7 +2,7 @@
 
 module ClaimsApi
   module V2
-    module PowerOfAttorneyRequest
+    module PowerOfAttorneyRequests
       class BaseController < ApplicationController
         # TODO: Figure out appropriate status codes for various upstream service
         # communication failure modes. Using:
