+++ modules/claims_api/app/models/claims_api/power_of_attorney_request.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+module ClaimsApi
+  class PowerOfAttorneyRequest <
+    Data.define(
+      :id,
+      :power_of_attorney_code,
+      :veteran,
+      :claimant,
+      :claimant_address,
+      :decision,
+      :authorizes_address_changing,
+      :authorizes_treatment_disclosure,
+      :created_at
+    )
+
+    extend Searching
+
+    Veteran =
+      Data.define(
+        :first_name,
+        :middle_name,
+        :last_name,
+        :participant_id
+      )
+
+    Claimant =
+      Data.define(
+        :first_name,
+        :last_name,
+        :participant_id,
+        :relationship_to_veteran
+      )
+
+    Address =
+      Data.define(
+        :city, :state, :zip, :country,
+        :military_post_office,
+        :military_postal_code
+      )
+  end
+end
