+++ modules/claims_api/app/models/claims_api/power_of_attorney_request/decision.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module ClaimsApi
+  class PowerOfAttorneyRequest
+    class Decision <
+      Data.define(
+        :status,
+        :representative,
+        :declined_reason,
+        :updated_at
+      )
+
+      module Statuses
+        ALL = [
+          NEW = 'New',
+          PENDING = 'Pending',
+          ACCEPTED = 'Accepted',
+          DECLINED = 'Declined'
+        ].freeze
+      end
+
+      Representative =
+        Data.define(
+          :first_name,
+          :last_name,
+          :email
+        )
+    end
+  end
+end
