+++ modules/claims_api/config/routes.rb
@@ -65,7 +65,7 @@ ClaimsApi::Engine.routes.draw do
     end
 
     resources :power_of_attorney_requests, path: 'power-of-attorney-requests', only: [:index] do
-      scope module: :power_of_attorney_request, param: :id do
+      scope module: :power_of_attorney_requests, param: :id do
         member do
           resource :decision, only: [:update]
         end
